requirements.txt
streamlit==1.44.1
pillow==11.1.0
torch==2.7.0
torchvision==0.22.0